+++ src/container/Types.ts
@@ -1,0 +1,9 @@
+export enum ContainerStatus {
+	STOPPED = 0,
+	STARTING = 1,
+	RUNNING = 2,
+	STOPPING = 3,
+	FAILED = 4,
+	INITIALIZING = 5,
+	ONLINE = 6
+};
